# docs/swagger.yaml
definitions:
...
  models.CreateUserRequest:
    properties:
      academicYear:
        type: string
      department:
        type: string
      email:
        maxLength: 100
        type: string
      fullName:
        maxLength: 100
        type: string
      lecturerId:
        type: string
      password:
        minLength: 6
        type: string
      programStudy:
        type: string
      roleId:
        type: string
      studentId:
        type: string
      username:
        maxLength: 50
        type: string
    required:
    - email
    - fullName
    - password
    - roleId
    - username
    type: object
...
  models.UpdateUserRequest:
    properties:
      email:
        maxLength: 100
        type: string
      fullName:
        maxLength: 100
        type: string
      isActive:
        type: boolean
      roleId:
        type: string
      username:
        maxLength: 50
        type: string
    type: object
  models.UpdateUserRoleRequest:
    properties:
      roleId:
        type: string
    required:
    - roleId
    type: object
...
  models.UserResponse:
    properties:
      email:
        type: string
      fullName:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      permissions:
        items:
          type: string
        type: array
      role:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
  description: API Server untuk manajemen dan pelaporan prestasi mahasiswa.
  title: Sistem Pelaporan Prestasi API
  version: "1.0"
paths:
...
  /api/v1/users:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of all registered users
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: List all users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a new user record in the system (Admin only)
      parameters:
      - description: User Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Email/Username already taken
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new user
      tags:
      - Users
  /api/v1/users/{id}:
    delete:
      consumes:
      - application/json
      description: Remove user from the system
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete user
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Retrieve detailed information of a specific user
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update user personal information
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      - description: Updated User Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update user
      tags:
      - Users
  /api/v1/users/{id}/role:
    put:
      consumes:
      - application/json
      description: Change the access role of a user
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      - description: Role Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User role updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update user role
      tags:
      - Users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"